+++ changelog.md
@@ -1,7 +1,9 @@
-v0.2.0
+v0.1.1-beta
 - Switched to RolflectionLib to stop relying on kotlin
 - Switched to WrapUI Dialogs for more consistent behaviour
 - Bug fixes
+- The sector now starts with stockpiles to prevent death spirals.
+- Larger markets now have a smaller ratio of their population as workers, while smaller markets more.
 
-v0.1.0
+v0.1.0-beta
 - Initial release